[frontend] improve platofrm alert dialog for genericity
