+++ opencti-platform/opencti-front/lang/fr.json
@@ -2530,5 +2530,6 @@
   "Open Settings": "Ouvrir les paramètres",
   "alert_USER_WITH_NULL_EFFECTIVE_LEVEL": "A partir d'OpenCTI 6.0, un niveau de confiance maximum sera ajouté aux utilisateurs. Il permettra aux administrateurs de la plateforme de mieux contrôler l'impact des données importées et d'éviter toute mise à jour indésirable des données. Veuillez lire {link_blogpost} pour plus d'informations.\n \nSur cette plateforme, le niveau de confiance maximum de certains utilisateurs est actuellement indéfini. Ils ne pourront pas créer de données dans la version 6.X de la plateforme. Pour éviter cela, assurez-vous que tous les utilisateurs ont un niveau de confiance maximal défini. Plus d'informations dans {link_blogpost}. Soyez particulièrement attentif aux utilisateurs associés aux connecteurs, aux flux et aux courants. Rejoignez {link_slack} si vous avez besoin de conseils.",
   "No confidence level found in this user's groups and organizations, and no confidence level defined at the user level.": "Aucun niveau de confiance n'a été trouvé dans les groupes et organisations de cet utilisateur, et aucun niveau de confiance n'a été défini au niveau de l'utilisateur.",
-  "confidence_level_from": "De {entity_type} {link}"
+  "confidence_level_from": "De {entity_type} {link}",
+  "Unknown configuration error in the platform.": "Erreur de configuration inconnue dans la plate-forme."
 }
